# app/src/features/scene-analysis/model.ts
import type { components } from "../../shared/api/schema";

export type Rectangle = components["schemas"]["Rectangle"];
export type Regions = components["schemas"]["Regions"];
export type Region = Regions["target"];
export type Result = components["schemas"]["Result"];
export type Metrics = components["schemas"]["Metrics"];
export type RegionName = "target" | "local_background" | "bright_background";
export type Selection = Partial<Record<RegionName, Rectangle>>;
export type CompleteSelection = Record<RegionName, Rectangle>;

export const REGION_NAMES: RegionName[] = ["target", "local_background", "bright_background"];

export const REGION_LABELS: Record<RegionName, string> = {
  target: "Target",
  local_background: "Local Background",
  bright_background: "Bright Background",
};

export function complete(selection: Selection): selection is CompleteSelection {
  return REGION_NAMES.every((name) => selection[name] !== undefined);
}